move carleton views list out of menus/tabs.js
