+++ views/menus/carleton-list.js
@@ -6,8 +6,61 @@ import type {CarletonDetailMenuType} from './types'
 import {Row} from '../components/layout'
 import {ListRow, ListSeparator, Title} from '../components/list'
 import {NoticeView} from '../components/notice'
+import {BonAppHostedMenu} from './menu-bonapp'
 
-import {carleton} from './tabs'
+
+const carleton = [
+  {
+    id: 'CarletonBurtonMenuView',
+    title: 'Burton',
+    rnVectorIcon: {iconName: 'trophy'},
+    component: BonAppHostedMenu,
+    props: {
+      cafeId: '35',
+      loadingMessage: [
+        'Searching for Schiller…',
+      ],
+    },
+  },
+  {
+    id: 'CarletonLDCMenuView',
+    title: 'LDC',
+    rnVectorIcon: {iconName: 'water'},
+    component: BonAppHostedMenu,
+    props: {
+      cafeId: '36',
+      loadingMessage: [
+        'Tracking down empty seats…',
+      ],
+    },
+  },
+  {
+    id: 'CarletonWeitzMenuView',
+    title: 'Weitz Center',
+    rnVectorIcon: {iconName: 'wine'},
+    component: BonAppHostedMenu,
+    props: {
+      cafeId: '458',
+      loadingMessage: [
+        'Observing the artwork…',
+        'Previewing performances…',
+      ],
+    },
+  },
+  {
+    id: 'CarletonSaylesMenuView',
+    title: 'Sayles Hill',
+    rnVectorIcon: {iconName: 'snow'},
+    component: BonAppHostedMenu,
+    props: {
+      cafeId: '34',
+      loadingMessage: [
+        'Engaging in people-watching…',
+        'Checking the mail…',
+      ],
+    },
+  },
+]
 
 const styles = StyleSheet.create({
   rowText: {
