+++ views/menus/tabs.js
@@ -64,58 +64,4 @@ const stolaf = [
   // },
 ]
 
-const carleton = [
-  {
-    id: 'CarletonBurtonMenuView',
-    title: 'Burton',
-    rnVectorIcon: {iconName: 'trophy'},
-    component: BonAppHostedMenu,
-    props: {
-      cafeId: '35',
-      loadingMessage: [
-        'Searching for Schiller…',
-      ],
-    },
-  },
-  {
-    id: 'CarletonLDCMenuView',
-    title: 'LDC',
-    rnVectorIcon: {iconName: 'water'},
-    component: BonAppHostedMenu,
-    props: {
-      cafeId: '36',
-      loadingMessage: [
-        'Tracking down empty seats…',
-      ],
-    },
-  },
-  {
-    id: 'CarletonWeitzMenuView',
-    title: 'Weitz Center',
-    rnVectorIcon: {iconName: 'wine'},
-    component: BonAppHostedMenu,
-    props: {
-      cafeId: '458',
-      loadingMessage: [
-        'Observing the artwork…',
-        'Previewing performances…',
-      ],
-    },
-  },
-  {
-    id: 'CarletonSaylesMenuView',
-    title: 'Sayles Hill',
-    rnVectorIcon: {iconName: 'snow'},
-    component: BonAppHostedMenu,
-    props: {
-      cafeId: '34',
-      loadingMessage: [
-        'Engaging in people-watching…',
-        'Checking the mail…',
-      ],
-    },
-  },
-]
-
-export {stolaf, carleton}
 export default stolaf
